requirements.txt
requests==2.31.0
finnhub-python==2.4.18
openai==1.3.0
pandas==2.1.3
numpy==1.26.2
beautifulsoup4==4.12.2
feedparser==6.0.10
tweepy==4.14.0
praw==7.7.1
textblob==0.17.1
vaderSentiment==3.3.2
yfinance==0.2.32
alpha-vantage==2.3.1
python-dotenv==1.0.0
schedule==1.2.0
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-cloud-logging==3.8.0
google-cloud-secret-manager==2.16.4
flask==3.0.0
gunicorn==21.2.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
